refactor: move least_square to curves submodule
